tests: Moves dummy skeleton & heights fixtures to tests/conftest.py
Closes #843

Fixtures generated under `tests/tracing/conftest.py` are now generated under `tests/conftest.py` so that they are
available for other tests (at this point `tests/measure/test_height.py` that is in development).
