+++ tests/conftest.py
@@ -9,7 +9,7 @@ import pandas as pd
 import pySPM
 import pytest
 import yaml
-from skimage import draw
+from skimage import draw, filters
 from skimage.morphology import skeletonize
 
 import topostats
@@ -759,3 +759,112 @@ def utils_skeleton_linear3() -> npt.NDArray:
     """Linear skeleton with several branches."""
     random_images, _ = draw.random_shapes(rng=7334281, **kwargs)
     return skeletonize(random_images != 255)
+
+
+# Pruning and Height profile fixtures
+#
+# Skeletons are generated by...
+#
+# 1. Generate random boolean images using scikit-image.
+# 2. Skeletonize these shapes (gives boolean skeletons), these are our targets
+# 3. Scale the skeletons by a factor (100)
+# 4. Apply Gaussian filter to blur the heights and give an example original im
+#    for.
+
+
+def _generate_heights(skeleton: npt.NDArray, scale: float = 100, sigma: float = 5.0, cval: float = 20.0) -> npt.NDArray:
+    """Generate heights from skeletons by scaling image and applying Gaussian blurring.
+
+    Uses scikit-image 'skimage.filters.gaussian()' to generate heights from skeletons.
+
+    Parameters
+    ----------
+    skeleton : npt.NDArray
+        Binary array of skeleton.
+    scale : float
+        Factor to scale heights by. Boolean arrays are 0/1 and so the factor will be the height of the skeleton ridge.
+    sigma : float
+        Standard deviation for Gaussian kernel passed to `skimage.filters.gaussian()'.
+    cval : float
+        Value to fill past edges of input, passed to `skimage.filters.gaussian()'.
+
+    Returns
+    -------
+    npt.NDArray
+        Array with heights of image based on skeleton which will be the backbone and target.
+    """
+    return filters.gaussian(skeleton * scale, sigma=sigma, cval=cval)
+
+
+def _generate_random_skeleton(**extra_kwargs):
+    """Generate random skeletons and heights using skimage.draw's random_shapes()."""
+    kwargs = {
+        "image_shape": (128, 128),
+        "max_shapes": 20,
+        "channel_axis": None,
+        "shape": None,
+        "allow_overlap": True,
+    }
+    heights = {"scale": 100, "sigma": 5.0, "cval": 20.0}
+    random_image, _ = draw.random_shapes(**kwargs, **extra_kwargs)
+    mask = random_image != 255
+    skeleton = skeletonize(mask)
+    return {"img": _generate_heights(skeleton, **heights), "skeleton": skeleton}
+
+
+@pytest.fixture()
+def skeleton_loop1() -> dict:
+    """Skeleton with loop to be retained and side-branches."""
+    return _generate_random_skeleton(rng=1, min_size=20)
+
+
+@pytest.fixture()
+def skeleton_loop2() -> dict:
+    """Skeleton with loop to be retained and side-branches."""
+    return _generate_random_skeleton(rng=165103, min_size=60)
+
+
+@pytest.fixture()
+def skeleton_linear1() -> dict:
+    """Linear skeleton with lots of large side-branches, some forked."""
+    return _generate_random_skeleton(rng=13588686514, min_size=20)
+
+
+@pytest.fixture()
+def skeleton_linear2() -> dict:
+    """Linear Skeleton with simple fork at one end."""
+    return _generate_random_skeleton(rng=21, min_size=20)
+
+
+@pytest.fixture()
+def skeleton_linear3() -> dict:
+    """Linear Skeletons (i.e. multiple) with branches."""
+    return _generate_random_skeleton(rng=894632511, min_size=20)
+
+
+## Helper function visualising for generating skeletons and heights
+
+# def pruned_plot(gen_shape: dict) -> None:
+#     """Plot the original skeleton, its derived height and the pruned skeleton."""
+#     img_skeleton = gen_shape()
+#     pruned = topostatsPrune(
+#         img_skeleton["heights"],
+#         img_skeleton["skeleton"],
+#         max_length=-1,
+#         height_threshold=90,
+#         method_values="min",
+#         method_outlier="abs",
+#     )
+#     pruned_skeleton = pruned._prune_by_length(pruned.skeleton, pruned.max_length)
+#     fig, ((ax1, ax2), (ax3, ax4)) = plt.subplots(2, 2)
+#     ax1.imshow(img_skeleton["skeleton"])
+#     ax2.imshow(img_skeleton["heights"])
+#     ax3.imshow(pruned_skeleton)
+#     plt.show()
+
+
+# pruned_plot(pruning_skeleton_loop1)
+# pruned_plot(pruning_skeleton_loop2)
+# pruned_plot(pruning_skeleton_linear1)
+# pruned_plot(pruning_skeleton_linear2)
+# pruned_plot(pruning_skeleton_linear3)
